Ajustes em Triagem e na NavBar
