+++ FRONT/src/components/inicio/styledInicio.jsx
